feat: pos renders separately on checkout page
